requirements.txt
absl-py==0.5.0
appnope==0.1.0
astor==0.7.1
astroid==2.0.4
backcall==0.1.0
Click==7.0
contextlib2==0.5.5
cycler==0.10.0
Cython==0.29
decorator==4.3.0
dlib==19.16.99
face-recognition==1.2.3
face-recognition-models==0.3.0
gast==0.2.0
grpcio==1.15.0
h5py==2.8.0
ipykernel==5.1.0
ipython==7.0.1
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.13.1
jupyter-client==5.2.3
jupyter-core==4.4.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lxml==4.2.5
Markdown==3.0.1
matplotlib==3.0.0
mccabe==0.6.1
numpy==1.15.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prompt-toolkit==2.0.5
protobuf==3.6.1
ptyprocess==0.6.0
pycocotools==2.0.0
Pygments==2.2.0
pylint==2.1.1
pyparsing==2.2.2
python-dateutil==2.7.3
pyzmq==17.1.2
simplegeneric==0.8.1
six==1.11.0
tensorboard==1.11.0
tensorflow==1.11.0
termcolor==1.1.0
tornado==5.1.1
traitlets==4.3.2
typed-ast==1.1.0
wcwidth==0.1.7
Werkzeug==0.14.1
wrapt==1.10.11
